Receive response from CivicsApiService.kt

+++ app/src/main/java/com/example/android/politicalpreparedness/ui/election/ElectionsFragment.kt
@@ -54,7 +54,7 @@ class ElectionsFragment: Fragment() {
         val adapter = ElectionListAdapter(ElectionListAdapter.ElectionListener { election ->
             findNavController().navigate(
                 ElectionsFragmentDirections.actionElectionsFragmentToVoterInfoFragment(
-                    election.id,
+                    election.id.toLong(),
                     election.division
                 )
             )
